Move url utils to separate file to keep entry clean
